PYG-14/PYG-68/Feat(repository-controller-service-html)/Verificacao-Regionalismo

+++ src/main/java/edu/fatec/Porygon/repository/SinonimoRepository.java
@@ -5,7 +5,7 @@ import edu.fatec.Porygon.model.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface SinonimoRepository extends JpaRepository<Sinonimo, Long> {
+public interface SinonimoRepository extends JpaRepository<Sinonimo, Integer> {
 
     boolean existsByNomeAndTag(String nome, Tag tag);
     void deleteByTag(Tag tag);
